rename Deployment Checker service
