+++ deployment-checker/docs/swagger/swagger.yaml
